requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
cffi==1.15.1
charset-normalizer==3.3.2
cycler==0.11.0
decorator==5.1.1
einops==0.3.2
fonttools==4.38.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.1.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.39.1
matplotlib==3.5.0
msgpack==1.0.5
natsort==8.0.0
numba==0.56.4
numpy==1.21.6
omegaconf==2.3.0
packaging==23.2
pesq==0.0.3
Pillow==9.5.0
platformdirs==3.11.0
pooch==1.8.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.2
setuptools-scm==7.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.10.0
torchaudio==0.10.0
torchinfo==1.5.3
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
